Improve reporter modularization and documentation
