+++ src/lib/mainEvent.spec.js
@@ -1,0 +1,3 @@
+it("should be a singleton", () => {
+  expect(require("./mainEvent.js")).toBe(require("./mainEvent.js"));
+});
